requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
certifi==2020.12.5
chardet==3.0.4
charset-normalizer==3.0.1
config==0.5.0.post0
frozenlist==1.3.3
idna==2.10
meraki==1.27.0
multidict==6.0.4
paho-mqtt==1.5.1
pymongo==3.11.2
requests==2.25.0
requests-toolbelt==0.9.1
urllib3==1.26.2
yarl==1.8.2
